Reach the end of the reporting flow in every branch
You can reach the end of the user-side reporting flow for every branch. TODO: Still need to save the info the user gives somewhere so that we can differentiate some branches and give something to the mod side. Also need to test this code more thoroughly to make sure it doesn't get stuck or crash no matter what the user enters and in what order.